requirements.txt
# AI Investment System - Python Dependencies
# Install: pip install -r requirements.txt

# Google Sheets API
gspread>=6.0.0
google-auth>=2.28.0

# Data & Computation
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0

# Market Data
yfinance>=0.2.36
requests>=2.31.0

# Dashboard (local development only)
streamlit>=1.30.0
